requirements.txt
imgcompare==2.0.1
pillow==11.3.0
tabulate==0.9.0
tqdm==4.67.1
